173 duplicates (#218)
* WIP - worker adjustments for multiple duplicates

* WIP - API adjustments for multiple duplicates

* fixed overwritten state alert

* added bundle cascade

* fixed accept

* changed button wording

* check dataset state before acceptance or rejection

* WIP - testing

* fixed alert

* persist action item instead of notification

* show if action item is resolved

* removed unused method

* update action item states; conditionally disable controls on action items

* removed onclick

* update action item status

* update version of rejected duplicates

* fixed alerts logic

* changed alerts logic

* fetch notifications by active action items

* trigger notification retrieval when action item is resolved

* set max width on notification dropdown

* updated workflow config for rejection

* moved writes to API layer

* refactored DB writes to API layer; added comment wrap plugin

* added comment

* use createMany with skipDuplicates

* simplified acceptance into resumable steps

* WIP - post duplicate by id

* refactoring API and worker code

* fixed URL; added max height

* fixed duplicate creation

* fixed state check

* fixed create action item

* added duplicate ready state

* include duplications in analysis

* use array

* allow acceptance to resume

* check paths before deletion

* update action items before resuming workflow

* ack notification before acceptance

* use upsert

* check exists before create

* fix user connect query

* added error for testing

* fixed reject endpoint call

* null check

* skip rejecting duplicates

* removed code for testing

* reverted try catch

* removed code for testing

* logging

* logging

* Revert "removed code for testing"

This reverts commit 092b7a26447a461579898429326048bc4eb5fb22.

* Revert "removed code for testing"

This reverts commit a19adccf160017e7545239f671817798019fb1ed.

* reverted rejection of other diplicates

* modified checks after preflight

* use separate checks pre-and post-accpetance

* broke acceptance into resumable chunks

* added logging

* fixed return

* fixed complete; state check

* use .state

* check correct state

* commented out error

* create new states

* fixed conditions for alerts

* fixed access

* null checks

* use is_deleted

* removed error

* enable resume

* update state once

* added workflow for reject

* fixed rejection

* logging

* logging

* logging

* logging

* added await

* check states before purging resources

* fixed state

* error for testing

* use separate states

* fix condition

* fixed condition

* changes test condition

* fixed condition

* show active duplicates

* removed test error

* max h

* show alerts

* spacing

* bold

* clear input

* font

* class danger

* hide default actions

* close modal

* refactored workflow

* notification ack fix

* fixed retrieval

* fixed returns

* changed to consider active action items

* use reject

* fixed returns

* logging

* active not.

* filter notifications

* fetch by status

* added list component for duplicates

* use new comp.

* v-if

* logging

* fix

* fixed cond.

* show deleted

* use right field

* label

* fixed state

* added state checks for API layer

* responsive

* responsive

* font

* fix

* fixed state check

* refresh on error

* WIP - lock actions

* disabled controls; cleanup

* disabled controls

* lock workers based on state

* disable deletions

* fixed await

* fixed imports

* fixes

* added state alert to DatasetDownloadModal.vue

* modal fix

* fetch in transaction

* settings.json

* docker compose nodemon

* error

* lock when state can'tbe determined

* assume locked

* no dismiss

* lock delete

* alerts

* duplication includes

* include duplications

* param

* fixed include

* fixed alerts

* button text

* alerts

* imports

* alerts

* reverted

* test errors

* check index before removal from list (#197)

* Bump express from 4.18.2 to 4.19.2 in /secure_download

Bumps [express](https://github.com/expressjs/express) from 4.18.2 to 4.19.2.
- [Release notes](https://github.com/expressjs/express/releases)
- [Changelog](https://github.com/expressjs/express/blob/master/History.md)
- [Commits](https://github.com/expressjs/express/compare/4.18.2...4.19.2)

---
updated-dependencies:
- dependency-name: express
  dependency-type: direct:production
...


* Bump express from 4.18.3 to 4.19.2 in /api

Bumps [express](https://github.com/expressjs/express) from 4.18.3 to 4.19.2.
- [Release notes](https://github.com/expressjs/express/releases)
- [Changelog](https://github.com/expressjs/express/blob/master/History.md)
- [Commits](https://github.com/expressjs/express/compare/4.18.3...4.19.2)

---
updated-dependencies:
- dependency-name: express
  dependency-type: direct:production
...


* Rewrite bundle population and download logic (#202)

* populate bundles

* testing resuming wf

* added logging

* logging

* logging

* pause existing wfs

* logging

* unstage datasets first

* testing resuming wf

* changed is_staged cond.

* removed error

* fixed condition

* logging

* logging

* delete pending wf

* delete tasks too

* refactoring

* fixed init

* fixed text

* cleanup

* use num_files check

* use bundle_alias check

* populate bundles and unstage files

* migration

* changed script name

---------


* test error

* use array

* disable notifications for users

* create project-dataset associations for duplicates; filter out duplicates form search

* allow modal dismiss

* removed redundant update

* removed test error

* omit duplicates

* omit duplicates

* logging

* omit deleted

* lock state

* loading

* descending

* Adjust docker compose for Dockerized secure_download API

* make bundle workflows agnostic of filesystem path (#206)

* make bundle workflows agnostic of filesystem path

* fixed docs

* testing

---------


* migration (#207)

* removed test error

* unrelated change

* removed error

* fixed utils

* removed breakpoint

* syntax fix

* use correct path

* use correct path

* allow file browsing

* disabled timeout

* use ref

* deleted file

* file browser alert

* syntax fix (#209)

* removed path (#210)


* refresh after download

* refactoring

* refactored duplication code into service

* imports

* use download step

* Revert "imports"

This reverts commit 9ab5beb936b5cb353c6608f049213341bb2b2193.

* Revert "refactored duplication code into service"

This reverts commit 79fe40b68fb8b5403b159f763bfe2966fe0ca975.

* remove state check from download

* added logging

* refresh dataset after download is initiated

* pass prop

* logging

* logging

* emit event from intermediate component

* logging

* logging

* catch event in table

* refresh upon file download

* check for roles

* remove role check from review

* check if ds is duplicate for write endpoints

* case

* some cleanup

* removed redundant state checks

* refactored into smaller modules

* separated into methods

* changed endpoints

* logging

* misc fixes

* fix concat

* moved states into constants

* renaming; docs

* cleanup

* renamed table

* added state diagram 1

* added more docs

* deleted migrations

* migration

* updated docs

* fix action item view

* remove version from UI

* updated docs

---------

+++ workers/workers/scripts/purge_staged_datasets.py
@@ -4,6 +4,7 @@ from pathlib import Path
 
 import workers.api as api
 from workers.config import config
+from workers.dataset import get_bundle_staged_path, is_dataset_locked_for_writes
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -25,9 +26,14 @@ def main():
         'staged_path': None
     }
     for dataset in datasets[:MAX_PURGES]:
+        locked, latest_state = is_dataset_locked_for_writes(dataset)
+        if locked:
+            logger.warning(f"Dataset {dataset['id']}'s current state is {latest_state}, and it is locked for writes. Dataset will not be purged.'")
+            continue
+
         try:
             staged_path = Path(dataset['staged_path'])
-            bundle_path = Path(f'{dataset["bundle"]["path"]}')
+            bundle_path = Path(get_bundle_staged_path(dataset=dataset))
 
             if staged_path.exists():
                 shutil.rmtree(staged_path)
